dockerized file, writing readme and testing
