+++ .dockerignore
@@ -1,0 +1,9 @@
+node_modules
+.next
+.git
+.env
+npm-debug.log*
+README.md
+.dockerignore
+Dockerfile
+docker-compose.yml
